\name{draw}
\alias{draw}

\title{Generic object drawing}
\description{Generic function for drawing objects.}


\usage{draw(x, ...)}


\arguments{
  \item{x}{the object to be drawn.}
  \item{...}{graphical arguments to be passed to S3 methods.}
}

\details{
This is a S3-style UseMethod dispatcher.
}
